# Conflicts:
#	Dev/Cpp/Effekseer/Effekseer/Effekseer.Server.cpp
#	Dev/Cpp/Effekseer/Effekseer/Effekseer.Server.h
#	Dev/Cpp/Effekseer/Effekseer/Effekseer.ServerImplemented.h
#	Dev/Cpp/Effekseer/Effekseer/Effekseer.Socket.cpp
#	Dev/Cpp/Effekseer/Effekseer/Effekseer.Socket.h

+++ Dev/Cpp/Effekseer/Effekseer/Culling/Culling3D.Grid.cpp
@@ -17,7 +17,7 @@ namespace Culling3D
 		assert(o_->ObjectIndex == -1);
 
 		objects.push_back(o_);
-		o_->ObjectIndex = objects.size() - 1;
+		o_->ObjectIndex = (int32_t)objects.size() - 1;
 	}
 
 	void Grid::RemoveObject(Object* o)
