+++ Dev/Cpp/Effekseer/Effekseer/Culling/Culling3D.Vector3DF.cpp
@@ -117,6 +117,6 @@ namespace Culling3D
 		float dx = v1.X - v2.X;
 		float dy = v1.Y - v2.Y;
 		float dz = v1.Z - v2.Z;
-		return sqrt(dx * dx + dy * dy + dz * dz);
+		return sqrtf(dx * dx + dy * dy + dz * dz);
 	}
 }
